# splinterd/api/static/openapi.yml
...
openapi: '3.0.0'
info:
  version: 0.3.12
  title: SplinterD API
  description: REST API for SplinterD
servers:
  - url: http://localhost:9000/api
paths:
...
  /biome/register:
    post:
      tags:
        - Biome
      description: Create new user with username and password credentials
      parameters:
        - $ref: "#/components/parameters/protocol_version"
      requestBody:
        content:
          application/json:
            schema:
              properties:
                username:
                  description: username of a new user
                hashed_password:
                  description: |
                     Hashed password to be used for user authentication
              required:
                - username
                - hashed_password
              example:
                username: alice@acme.com
                hashed_password: F4AABE0B40C9ABB8B6FD2EEACB39C965
      responses:
        200:
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User created successfully"
        400:
          description: Invalid request
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/ErrorBiome'
        500:
          description: Internal server error occurred
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/ErrorBiome'
  /biome/login:
    post:
      tags:
        - Biome
      description: Authenticates a user with username and password credentials
      parameters:
        - $ref: "#/components/parameters/protocol_version"
      requestBody:
        content:
          application/json:
            schema:
              properties:
                username:
                  description: username of user
                hashed_password:
                  description: |
                    Hashed password to be used for user authentication

              required:
                - username
                - hashed_password
              example:
                username: alice@acme.com
                hashed_password: |-
                  8945622435187243046536949706b5272644c71336c7254563679727565494b376d4b3554696b734662685962652f6v52562e437a70462f6552489c8b
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Successful login"
                  user_id:
                    type: string
                    description: "Internal unique identifier for the user"
                    example: "f35aacc1-a9cd-4eda-b6d0-2efaddf0c8a4"
                  token:
                    type: string
                    description: "JWT access token used for authorizing access to protected resources"
                    example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lk\
                      IjoiZjM1YWFjYzEtYTljZC00ZWRhLWI2ZDAtMmVmYWRkZjBjOGE0Iiwia\
                      XNzIjoic2VsZi1pc3N1ZWQiLCJleHAiOjE1ODAyMzkyMjh9.P8hA0ru_x\
                      riYX7qryl08ZEp86t5HD_AEVPEUXY70Ehc"
        400:
          description: Invalid request
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/ErrorBiome'
        500:
          description: Internal server error occurred
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/ErrorBiome'
  /biome/users:
    get:
      tags:
        - Biome
      description: Lists all users
      parameters:
        - $ref: "#/components/parameters/protocol_version"
      responses:
        200:
          description: List of users registered in Biome
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    username:
                      type: string
                      description: "Username used for login"
                      example: "alice@acme.com"
                    user_id:
                      type: string
                      description: "Internal unique identifier for the user"
                      example: "f35aacc1-a9cd-4eda-b6d0-2efaddf0c8a4"
        500:
          description: Internal server error occurred
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/ErrorBiome'
  /biome/users/{user_id}:
    get:
      tags:
      - Biome
      description: Fetch a user by ID
      parameters:
        - $ref: "#/components/parameters/protocol_version"
        - name: user_id
          in: path
          description: ID of the user
          required: true
          schema:
            type: string
            example: "f35aacc1-a9cd-4eda-b6d0-2efaddf0c8a4"
      responses:
        200:
          description: User information
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    description: "Username used for login"
                    example: "alice@acme.com"
                  user_id:
                    type: string
                    description: "Internal unique identifier for the user"
                    example: "f35aacc1-a9cd-4eda-b6d0-2efaddf0c8a4"
        400:
          description: Invalid request
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/ErrorBiome'
        404:
          description: Resource not found
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/ErrorBiome'
        500:
          description: Internal server error occurred
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/ErrorBiome'
    put:
      tags:
      - Biome
      description: Update a user
      parameters:
        - $ref: "#/components/parameters/protocol_version"
        - name: user_id
          in: path
          description: ID of the user
          required: true
          schema:
            type: string
            example: "f35aacc1-a9cd-4eda-b6d0-2efaddf0c8a4"
      requestBody:
        content:
          application/json:
            schema:
              properties:
                username:
                  description: Existing username of user
                hashed_password:
                  description: |
                    Hashed password to be used for user authentication
                new_password:
                  description: |
                    Hash of the new password to be used for user authentication
              required:
                - username
                - hashed_password
              example:
                username: alice@acme.com
                hashed_password: |-
                  8945622435187243046536949706b5272644c71336c7254563679727565494b376d4b3554696b734662685962652f6v52562e437a70462f6552489c8b
      responses:
        200:
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User updated successfully"
        400:
          description: Invalid request
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/ErrorBiome'
        404:
          description: Resource not found
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/ErrorBiome'
        500:
          description: Internal server error occurred
          content:
            application/json:
                schema:
                  $ref: '#/components/schemas/ErrorBiome'
...
components:
  parameters:
    protocol_version:
      schema:
        type: integer
        example: 1
      description: The protocol version which the client can understand
      in: header
      name: SplinterProtocolVersion
  schemas:
...
    ErrorBiome:
      additionalProperties: false
      properties:
        code:
          description: Error code
          type: string
          example: 400
        message:
          description: A message describing the error that occurred
          type: string
          example: "Invalid payload"
      required:
        - message
...